refactor: 롤링 js 파일분리, 함수 나누기
