+++ public/css/style.css
@@ -44,7 +44,6 @@
 }
 .front .search-keyword__bundle {
   position: absolute;
-  top: -165px;
 }
 .front .search-keyword__list {
   font-size: 25px;
