+++ public/index.html
@@ -8,6 +8,7 @@
     <script src="https://kit.fontawesome.com/c3d44d5c03.js" crossorigin="anonymous"></script>
     <script type="module" src="src/main.js" defer></script>
     <script type="module" src="src/getImg.js" defer></script>
+    <script type="module" src="src/getTextData.js" defer></script>
     <script type="module" src="src/carousel.js" defer></script>
     <script type="module" src="src/bottomCarousel.js" defer></script>
     <script type="module" src="src/rolling.js" defer></script>
@@ -20,7 +21,7 @@
                 <img src="images/kakocommerce.png" alt="" class="logo__img">
             </div>
             <div class="front__input-box">
-                <input type="text" placeholder="4 소파패드" class="input-box__input">
+                <input type="text" class="input-box__input">
             </div>
             <div class="front__search-keyword">
             </div>
