+++ public/sass/style.scss
@@ -68,7 +68,8 @@ $title_padding: 40px 300px;
   }
   .search-keyword__bundle {
     position: absolute;
-    top: -165px;
+    // top: -165px;
+    // top: 2px;
   }
 
   .search-keyword__list {
