+++ public/src/getImg.js
@@ -1,7 +1,7 @@
 import {loadBannerData} from './main.js';
 import {loadHomeContentsData} from './main.js';
-import {loadKeyword} from './main.js';
-import {staticContainer, slideContainer, slideBundle, issueContainer, issueBundle, keywordRollingDiv} from './main.js';
+
+import {staticContainer, slideBundle, issueContainer, issueBundle} from './main.js';
 const moreBtn = document.querySelector('.sub-menu__more');
 const subSlide = document.querySelector('.sub-menu__silde');
 function displayStaticImg(images) {
@@ -57,39 +57,16 @@ function displayissue(images) {
     }
     issueContainer.appendChild(issueBundle);
 }
-function displayKeywordRolling(keywords) {
-    let keywordsInfoArr = keywords.map(e => ({title: e.title, rank: e.rank}))
-    console.log(keywordsInfoArr);
-    let ul = document.createElement("ul");
-    ul.className = "search-keyword__bundle";
-    for(let i=0; i<10; i++) {
-        let li = document.createElement("li");
-        let liInnerText = document.createTextNode(`${keywordsInfoArr[i].rank}. ${keywordsInfoArr[i].title}`);
-        li.className = "search-keyword__list";
-        li.appendChild(liInnerText);
-        ul.appendChild(li);
-    }
-    keywordRollingDiv.appendChild(ul);
-    setInterval(function() {
-    ul.classList.toggle('move');
-    setTimeout(() => {
-        ul.appendChild(ul.firstElementChild);
-        ul.classList.toggle('move');
-    }, 300);
-    }, 2000);
-}
+
 let cnt =1;
 let cntSum=5;
 let cnt2 =1;
 function displaymoreContents(images) {
     cnt2++;
     let imageInfoArr = images.map( e => ({imgurl : e.eventContent.imgurl, title : e.eventContent.title }));
-    console.log(imageInfoArr);
     let div = document.createElement('div');
     div.className = "divBox";
     let ul;
-    console.log(cnt);
-    console.log(cntSum);
 for(let i=cnt; i<=cntSum; i++) {
     ul = document.createElement("ul")
     let liImg = document.createElement("li");
@@ -123,11 +100,9 @@ let cnt3 =0;
 function getMoreEvent(images) {
     cnt3++;
     moreBtn.removeEventListener('click', () => {
-        console.log('a');
         return displaymoreContents(images);
     });
     moreBtn.addEventListener('click', () => {
-        console.log('a');
         return displaymoreContents(images);
     })
 }
@@ -136,7 +111,6 @@ function getMoreEvent(images) {
 
 loadBannerData()
     .then(images => {
-        console.log(images);
         displayStaticImg(images);
         displaySlideImg(images);
     })
@@ -148,10 +122,4 @@ loadHomeContentsData()
         getMoreEvent(images);
     })
 
-loadKeyword()
-.then(keywords => {
-    // displayissue(images);
-    displayKeywordRolling(keywords);
-    // getMoreEvent(images);
-})
