+++ public/src/main.js
@@ -1,4 +1,3 @@
-// const mainSlide = document.querySelector(".main-banner__slide");
 const staticContainer = document.querySelector('.main-banner__static');
 const slideContainer = document.querySelector('.main-banner__slide');
 const slideBundle = document.querySelector(".slide-bundle");
@@ -6,7 +5,6 @@ const issueContainer = document.querySelector('.issue-banner__slide');
 const issueBanner = document.querySelector('.issue-banner');
 const issueBundle = document.querySelector('.issue-bundle');
 const keywordRollingDiv = document.querySelector('.front__search-keyword');
-// const keywordBundle = document.querySelector('.search-keyword__bundle');
